requirements.txt
pillow==8.0.1
numpy==1.19.4
requests==2.24.0
tqdm==4.51.0
matplotlib==3.3.2
fonttools==4.16.1
pandas==1.1.3
scipy==1.5.2
numba==0.51.2
wget==3.2
pypinyin==0.39.1
freetype-py==2.2.0
opencv-python>=4.4.0
beautifulsoup4>=4.6.0
imgaug==0.4.0
randomcolor
colormath
torch==1.8.1
torchvision==0.9.1
higher
wandb
scikit-image==0.17.2